+++ examples/using react and other custom scripts and styles/api.sql
@@ -1,0 +1,7 @@
+-- Adds a new entry in the clicks table
+INSERT INTO clicks(click_time) VALUES (datetime('now'));
+
+SELECT 'json' AS component,
+        JSON_OBJECT(
+            'total_clicks', (SELECT count(*) FROM clicks)
+        ) AS contents;