+++ examples/using react and other custom scripts and styles/my_react_component.js
@@ -5,8 +5,17 @@ function MyComponent({ greeting_name }) {
     const [count, setCount] = React.useState(0);
     return React.createElement(
         'button',
-        { onClick: () => setCount(count + 1), className: 'btn btn-primary' },
-        `Hello ${greeting_name}, you clicked me ${count} times!`
+        {
+            onClick: async () => {
+                const r = await fetch('/api.sql');
+                const { total_clicks } = await r.json();
+                setCount(total_clicks)
+            },
+            className: 'btn btn-primary'
+        },
+        count == 0
+            ? `Hello, ${greeting_name}. Click me !`
+            : `You clicked me ${count} times!`
     );
 }