+++ examples/using react and other custom scripts and styles/sqlpage/migrations/0001_clicks.sql
@@ -1,0 +1,3 @@
+CREATE TABLE clicks(
+    click_time TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+);