+++ src/webserver/http.rs
@@ -170,9 +170,8 @@ async fn build_response_header_and_stream<S: Stream<Item = DbItem>>(
                         database_entries_stream: stream,
                     });
                 }
-                PageContext::Close(h) => {
-                    head_context = h;
-                    break;
+                PageContext::Close(http_response) => {
+                    return Ok(ResponseWithWriter::FinishedResponse { http_response })
                 }
             },
             DbItem::FinishedQuery => log::debug!("finished query"),